Migrate to ViewPager2
